pyproject.toml
[project]
name = "jarvis-content-generator"
version = "1.0.0"
description = "AI Content Generator from Personal Knowledge Library"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.0.0",
    "anthropic>=0.39.0",
    "firecrawl-py>=1.0.0",
]

[project.scripts]
app = "content_app:app"

[tool.vercel]
# Optional: build step (none needed for Flask)
# build = "echo 'No build step'"
